Distinguish core state, reducer, and selectors
