+++ tensorboard/webapp/app.component.ts
@@ -14,7 +14,6 @@ limitations under the License.
 ==============================================================================*/
 import {Component, OnInit} from '@angular/core';
 import {Store} from '@ngrx/store';
-import {State} from './core/store';
 import {coreLoaded} from './core/actions';
 
 /** @typehack */ import * as _typeHackRxjs from 'rxjs';
@@ -25,7 +24,7 @@ import {coreLoaded} from './core/actions';
   styleUrls: ['./app.component.css'],
 })
 export class AppComponent implements OnInit {
-  constructor(private readonly store: Store<State>) {}
+  constructor(private readonly store: Store<{}>) {}
 
   ngOnInit() {
     this.store.dispatch(coreLoaded());
