+++ tensorboard/webapp/core/store/index.ts
@@ -14,3 +14,4 @@ limitations under the License.
 ==============================================================================*/
 
 export * from './core.reducers';
+export * from './core.selectors';
